Add date format utilitary functions
